+++ src/components/Layout/Nav/Nav.html
@@ -1,0 +1,14 @@
+<style>
+  nav {}
+
+  ul {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+  }
+</style>
+<nav>
+  <ul>
+    <slot name="item"></slot>
+  </ul>
+</nav>
