+++ src/pages/ChatPage/ChatPage.html
@@ -92,7 +92,13 @@
   </ypr-header>
 
   <ypr-main slot="chatsettings-main">
-    <ypr-nav></ypr-nav>
+    <ypr-nav>
+      <li slot="item"><a href="/signin">sign in</a></li>
+      <li slot="item"><a href="/signup">sign up</a></li>
+      <li slot="item"><a href="/error404">404</a></li>
+      <li slot="item"><a href="/error500">500</a></li>
+      <li slot="item"><a href="/dev">dev page</a></li>
+    </ypr-nav>
   </ypr-main>
 
   <ypr-footer slot="chatsettings-footer">
